chore: add subtitle directory and move to subtitle class
